added some code to the flow of the product odel

+++ Models/ProductData.cs
@@ -8,7 +8,7 @@ namespace Models
         {
 
         }
-        public ProductData(string name, string description, double price, int quantity)
+        public ProductData(string name, string description, decimal price, int quantity)
         {
             
             Name = name;
@@ -21,7 +21,7 @@ namespace Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
 
     }
